# swagger.yaml
openapi: 3.0.0
info:
  title: Car Contract Management System API
  description: 중고차 계약 관리 시스템 API 문서
  version: 1.0.0
  contact:
    name: API Support
    email: support@carmate.com
servers:
  - url: http://localhost:${process.env.PORT || 3000}
    description: Development server
    # 배포 전이므로 주석 처리
  # - url: ${process.env.PROD_SERVER_URL || 'https://api.carmate.com'}
  #  description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /users:
    post:
      tags:
        - Users
      summary: 회원가입
      description: 새로운 사용자 등록
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /users/{userId}:
    delete:
      tags:
        - Users
      summary: 사용자 삭제 (Admin)
      description: 특정 사용자 삭제 (관리자 권한 필요)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 사용자 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '유저 삭제 성공'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /companies:
    post:
      tags:
        - Companies
      summary: 회사 등록 (Admin)
      description: 새로운 회사 등록 (관리자 권한 필요)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: 회사 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - Companies
      summary: 회사 목록 조회 (Admin)
      description: 등록된 회사 목록 조회 (관리자 권한 필요)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - in: query
          name: searchBy
          schema:
            type: string
            enum: [companyName, companyCode]
        - $ref: '#/components/parameters/KeywordParam'
      responses:
        '200':
          description: 회사 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /companies/{companyId}:
    patch:
      tags:
        - Companies
      summary: 회사 정보 수정 (Admin)
      description: 회사 정보 수정 (관리자 권한 필요)
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyRequest'
      responses:
        '200':
          description: 회사 정보 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Companies
      summary: 회사 삭제 (Admin)
      description: 회사 삭제 (관리자 권한 필요)
      parameters:
        - in: path
          name: companyId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 회사 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '회사 삭제 성공'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /companies/users:
    get:
      tags:
        - Companies
      summary: 회사별 유저 목록 조회 (Admin)
      description: 회사별 유저 목록 조회 (관리자 권한 필요)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - in: query
          name: searchBy
          schema:
            type: string
            enum: [companyName, name, email]
        - $ref: '#/components/parameters/KeywordParam'
      responses:
        '200':
          description: 유저 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        default: 1
      description: 페이지 번호
    PageSizeParam:
      in: query
      name: pageSize
      schema:
        type: integer
        default: 8
      description: 페이지당 아이템 수
    KeywordParam:
      in: query
      name: keyword
      schema:
        type: string
      description: 검색 키워드
  schemas:
...
    RegisterUserRequest:
      type: object
      required:
        - name
        - email
        - employeeNumber
        - phoneNumber
        - password
        - passwordConfirmation
        - company
        - companyCode
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        employeeNumber:
          type: string
        phoneNumber:
          type: string
        password:
          type: string
          format: password
        passwordConfirmation:
          type: string
          format: password
        company:
          type: string
        companyCode:
          type: string
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        employeeNumber:
          type: string
        phoneNumber:
          type: string
        imageUrl:
          type: string
        isAdmin:
          type: boolean
        company:
          type: object
          properties:
            companyCode:
              type: string
    UserListResponse:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItemCount:
          type: integer
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              email:
                type: string
              employeeNumber:
                type: string
              phoneNumber:
                type: string
              company:
                type: object
                properties:
                  companyName:
                    type: string

    # Company schemas
    CreateCompanyRequest:
      type: object
      required:
        - companyName
        - companyCode
      properties:
        companyName:
          type: string
        companyCode:
          type: string
    UpdateCompanyRequest:
      type: object
      properties:
        companyName:
          type: string
        companyCode:
          type: string
    CompanyResponse:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        companyCode:
          type: string
        userCount:
          type: integer
    CompanyListResponse:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItemCount:
          type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/CompanyResponse'

    # Car schemas
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '잘못된 요청입니다'
            code: 'BAD_REQUEST'
            timestamp: '2024-01-01T00:00:00Z'
    Unauthorized:
      description: 인증이 필요합니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '로그인이 필요합니다'
            code: 'UNAUTHORIZED'
            timestamp: '2024-01-01T00:00:00Z'
...
    NotFound:
      description: 리소스를 찾을 수 없습니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '존재하지 않는 리소스입니다'
            code: 'NOT_FOUND'
            timestamp: '2024-01-01T00:00:00Z'
    Conflict:
      description: 충돌이 발생했습니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '이미 존재하는 데이터입니다'
            code: 'CONFLICT'
            timestamp: '2024-01-01T00:00:00Z'